agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "TripSync FastAPI backend for AI-powered trip planning"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.23.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "python-jose[cryptography]>=3.3.0",
    "supabase>=2.11.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
